backend/requirements.txt
boto3==1.34.123
chromadb==0.5.0
colorlog==6.8.2
Django==5.0.6
django-cors-headers==4.3.1
docx==0.2.4
docx2txt==0.8
langchain==0.2.3
langchain_core==0.2.5
langchain_google_vertexai==1.0.5
pandas==2.2.2
PyJWT==2.8.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
Requests==2.32.3
sentence_transformers==3.0.1
sqlparse==0.5.0
striprtf==0.0.26
tqdm==4.66.4
tqdm==4.66.4
typeguard==4.3.0
